Create Library & CLI
You can now use it in library mode and CLI mode!

The CLI should use the library and only parse the CLI, this means that this should be 100% library first with CLI as a second option.

This allows us to do more things in the future like a nice API :)