feat: new bfs for exiting early (#42)
We want to exit early if we have successfully decoded the text, this means we should stop calling other decoders after getting success as soon as possible!

To achieve this, without using for loops and having par_iter, we call try_for_each_with instead of map . we use channels to get our results. If we get success, we return None which attempts to stop the iterator ASAP!

 self.components
            .into_par_iter()
            .try_for_each_with(sender, |s, i| {
                let results = i.crack(text, &checker);
                if results.success {
                    s.send(results).expect("expected no send error!");
                    // returning None short-circuits the iterator
                    // we don't process any further as we got success
                    return None;
                }
                s.send(results).expect("expected no send error!");
                // return Some(()) to indicate that continue processing
                Some(())
            });
We now stop calling .crack() on other decoders when we get success, but it would be great if our bfs can exit early too right?

Therefore we use our custom results MyResult

pub enum MyResults {
    // return this if sucess
    Break(CrackResult),
    // else return all crack results
    Continue(Vec<CrackResult>),
}
And then inside bfs, if we encounter MyResult::Break, we know we have successfully decoded the text, so we use the same logic of try_for_each to short-circuit the iterator and exit early.

current_strings
            .into_iter()
            .map(|current_string| super::perform_decoding(&current_string))
            .try_for_each(|elem| match elem {
                // if it's Break variant, we have cracked the text successfully
                // so just stop processing further.
                MyResults::Break(res) => {
                    exit_result = Some(res);
                    None // short-circuits the iterator
                }
                MyResults::Continue(results_vec) => {
                    new_strings = results_vec
                        .into_iter()
                        .flat_map(|r| r.unencrypted_text)
                        .filter(|s| seen_strings.insert(s.clone()))
                        .collect();
                    Some(()) // indicate we want to continue processing
                }
            });
This also means that we won't have to iterate over all results again to see if any of them was successful!