Change positioning of timer::pause in human checker
Previously our logic looked like this:
* Check to see if API is on, if it is return True
* If it's not, pause human checker
* Perform human checker

This led to a glitch where the program would return with something, but the timer would still run and return a panic.

As we always return True for the HumanChecker if our API is on, we do not need to worry about whether or not the data passes any human checks and can just pause it.