Ensure BFS does not search on texts < 2 chars long
Previously Ares was trying to decode text with 1 or 2 chars:

```
[2022-12-12T13:01:16Z TRACE ares::decoders::binary_decoder] Trying binary with text "J"
```

I do not believe 1 or 2 chars are meaningful and they will not pass any of our checkers, so it is a bit pointless currently to try to decode them.

This PR adds a function which checks their size and if they are not large enough (more than 3 chars) they are not included in the strings to be decoded.

This should also provide a mild speed boost.